+++ src/Engine/RendererBase.h
@@ -5,13 +5,21 @@
 #include "SDL2/SDL.h"
 #include "Vector2.h"
 
+enum PrecacheStatus
+{
+	PRECACHE_SUCCESS = 1,
+	PRECACHEFAIL_UNKNOWN = 0,
+	PRECACHEFAIL_FAILTOLOAD = -1,
+	PRECACHEFAIL_CREATION = -2
+};
+
 class RendererBase
 {
 public:
 	virtual void Initialize();
 	virtual void Cleanup();
 
-	virtual void PrecacheTexture(const std::string& name, const std::string& path);
+	virtual int PrecacheTexture(const std::string& name, const std::string& path);
 	virtual void RenderImage(const char* textureID, Vector2 worldposition);
 	virtual void RenderImage(const char* textureID, Vector2 worldposition, double scale, double rotation, SDL_RendererFlip flip);
 	virtual void RenderDrawRect(Vector2 worldposition, Vector2 dimensions, SDL_Color color);
@@ -21,7 +29,7 @@ public:
 	virtual bool TextureExists(const std::string& Texture);
 
 	// Font Stuff
-	virtual void PrecacheFont(const std::string& name, const std::string& path, int fontsize);
+	virtual int PrecacheFont(const std::string& name, const std::string& path, int fontsize);
 	virtual void RenderFont(const std::string& message, const char* fontID, Vector2 worldposition, SDL_Color color);
 	virtual void RenderFont(const std::string& message, const char* fontID, Vector2 worldposition, double rotation, SDL_RendererFlip flip, SDL_Color color);